requirements.txt
pandas
numpy
scikit-learn
hydra-core
skorch
matplotlib
seaborn
hydra_colorlog